+++ compilers/hybrid-backend/src/utils/interval.cc
@@ -452,8 +452,8 @@ MultidimensionalIntervalSeq::~MultidimensionalIntervalSeq() {
         }
 }
 
-int MultidimensionalIntervalSeq::getNumOfElements() {
-	int count = intervals[0]->getNumOfElements();
+long int MultidimensionalIntervalSeq::getNumOfElements() {
+	long int count = intervals[0]->getNumOfElements();
 	for (int i = 1; i < dimensionality; i++) {
 		count *= intervals[i]->getNumOfElements();
 	}
